requirements(9).txt
--extra-index-url https://download.pytorch.org/whl/cu118
numpy==1.26.4
torch==2.1.0+cu118
torchvision==0.16.0+cu118
onnxruntime-gpu==1.16.3
gradio==5.9.1
opencv-python-headless==4.10.0.84
onnx==1.16.1
insightface==0.7.3
albucore==0.0.16
psutil==5.9.6
tqdm==4.66.4
ftfy
regex
pyvirtualcam
moviepy
gfpgan==1.3.8
mediapipe
safetensors>=0.3.1
realesrgan==0.3.0
albumentations>=1.1.0
tensorflow<2.14.0
transformers>=4.29.2
accelerate==0.23.0
